+++ src/main/java/seedu/ibook/logic/parser/IBookParser.java
@@ -13,15 +13,15 @@ import seedu.ibook.logic.commands.RedoCommand;
 import seedu.ibook.logic.commands.UndoCommand;
 import seedu.ibook.logic.commands.item.AddItemCommand;
 import seedu.ibook.logic.commands.item.DeleteItemCommand;
+import seedu.ibook.logic.commands.item.ExpiredCommand;
+import seedu.ibook.logic.commands.item.RemindCommand;
 import seedu.ibook.logic.commands.item.UpdateItemCommand;
 import seedu.ibook.logic.commands.product.AddCommand;
 import seedu.ibook.logic.commands.product.DeleteAllCommand;
 import seedu.ibook.logic.commands.product.DeleteCommand;
-import seedu.ibook.logic.commands.product.ExpiredCommand;
 import seedu.ibook.logic.commands.product.FindCommand;
 import seedu.ibook.logic.commands.product.ListCommand;
 import seedu.ibook.logic.commands.product.OutOfStockCommand;
-import seedu.ibook.logic.commands.product.RemindCommand;
 import seedu.ibook.logic.commands.product.UpdateAllCommand;
 import seedu.ibook.logic.commands.product.UpdateCommand;
 import seedu.ibook.logic.parser.exceptions.ParseException;
