+++ src/test/java/seedu/ibook/logic/commands/product/RemindCommandTest.java
@@ -18,6 +18,7 @@ import java.util.List;
 
 import org.junit.jupiter.api.Test;
 
+import seedu.ibook.logic.commands.item.RemindCommand;
 import seedu.ibook.model.Model;
 import seedu.ibook.model.ModelManager;
 import seedu.ibook.model.UserPrefs;
