+++ src/test/java/seedu/ibook/logic/parser/IBookParserTest.java
@@ -20,10 +20,10 @@ import seedu.ibook.logic.commands.ClearCommand;
 import seedu.ibook.logic.commands.ExitCommand;
 import seedu.ibook.logic.commands.RedoCommand;
 import seedu.ibook.logic.commands.UndoCommand;
+import seedu.ibook.logic.commands.item.ExpiredCommand;
 import seedu.ibook.logic.commands.product.AddCommand;
 import seedu.ibook.logic.commands.product.DeleteAllCommand;
 import seedu.ibook.logic.commands.product.DeleteCommand;
-import seedu.ibook.logic.commands.product.ExpiredCommand;
 import seedu.ibook.logic.commands.product.FindCommand;
 import seedu.ibook.logic.commands.product.ListCommand;
 import seedu.ibook.logic.commands.product.OutOfStockCommand;
